util: include predefined definitions during extraction
Effort to make regenerating files from kernel a little bit
easier including policy entries from separate file if it exists.
The policies in the kernel are mostly for netlink attributes used
from user-space to kernel. This allows us to specify kernel to
user-space attribute policies or override ones that pycparser has
a hard time with.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>