py80211: lib: get rid of python classes from netlink.core
The netlink.core module provides three native python classes which
wrap around the swig equivalent, ie. Callback, Socket, Message. However,
it is more of an inconvenience as more often as not there is need to
access the swig object. So lets leave those classes alone and stick with
the swig objects in py80211.

Signed-off-by: Arend van Spriel <aspriel@gmail.com>